+++ CHANGES.md
@@ -4,6 +4,7 @@
 ### `<v-grid>`
 - Drag selection is now disabled.
 - Row editor is now disabled.
+- `Value Generator` is now removed because of feature overlaps with `Renderer`
 - Issues fixed:
   - Grid doesn't work when using selection-mode multi, frozen columns and sortable columns. (#7)
   - Grid height is recalculated incorrectly when sorting a grid with a fixed height. (#8)